advanced_ai_agents/multi_agent_apps/agent_teams/ai_finance_agent_team/requirements_crewai.txt
crewai
crewai-tools
python-dotenv
openai
yfinance
